debug

+++ task.py
@@ -11983,7 +11983,7 @@ def train_L2W_v2(config,
           # reset model parameters
           weight_reset(snapshots)
           optimizer.iterations.assign(saved_step)
-        domain_rewards.assign(tf.constant(rewards))
+        domain_rewards.assign(tf.cast(tf.constant(rewards), dtype=domain_rewards.dtype))
         if not config.get("cosine_reward",True):
           domain_rewards.assign(tf.clip_by_value(domain_rewards, clip_value_min=-1.0, clip_value_max=1.0))
         # compute new domain distribution
